# openapi.yaml
openapi: 3.1.0
info:
  title: Synoptic Data Generation API
  version: 1.0.0
  description: API for managing synthetic data generation projects and jobs.
servers:
  - url: /api/v1 # Base path for API endpoints relative to the backend server URL
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A short error code or key.
        message:
          type: string
          description: A human-readable error message.
        details:
          type: object
          additionalProperties: true
          description: Optional additional error details.
      required:
        - error
        - message
...
    JobStatusValue:
      type: string
      enum: [queued, running, completed, failed, cancelled, paused, pending, accepted, rejected]
      description: The current status of the data generation job.
    JobProgress:
      type: object
      properties:
        percentComplete:
          type: number
          format: float
          minimum: 0
          maximum: 100
        currentStep:
          type: string
          description: Description of the current processing step.
        startTime:
          type: string
          format: date-time
          description: Timestamp when the job started processing.
        endTime:
          type: string
          format: date-time
          description: Timestamp when the job finished (completed, failed, or cancelled).
      required:
        - percentComplete
    JobError:
      type: object
      properties:
        code:
          type: string
          description: An error code identifying the type of failure.
        message:
          type: string
          description: A human-readable error message.
        details:
          type: object
          additionalProperties: true
          description: Optional additional error details.
      required:
        - code
        - message
    JobInputDataset:
      type: object
      properties:
        storageUri:
          type: string
          format: uri
          description: GCS URI to the input dataset.
        format:
          type: string
          description: Format of the input data (e.g., 'csv', 'json').
      # Make optional
    JobParameters:
      type: object
      properties:
        dataType:
          type: string
          description: Type of synthetic data to generate (e.g., 'customer_profiles', 'transactions').
        recordCount:
          type: integer
          format: int64
          description: Number of records to generate.
        format:
          type: string
          enum: [csv, json, parquet]
          description: Desired output format.
        quality:
          type: string
          enum: [draft, production]
          description: Desired quality level of the generated data.
        schemaDefinition: # Renamed from 'schema' to avoid conflict
          type: object
          description: Definition of the data schema to generate.
          additionalProperties: true # Allow flexible schema definition
        # Add other specific parameters needed for generation here
      required:
        - dataType
        - recordCount
        - format
        - quality
        - schemaDefinition
    JobOutputConfig:
      type: object
      properties:
        destinationUri: # Renamed from 'destination' for clarity
          type: string
          format: uri
          description: GCS URI where the output data should be written (within the project bucket).
      required:
        - destinationUri
    JobConfiguration:
      type: object
      properties:
        inputDataset:
          $ref: '#/components/schemas/JobInputDataset'
        parameters:
          $ref: '#/components/schemas/JobParameters'
        outputConfig:
          $ref: '#/components/schemas/JobOutputConfig'
      required:
        - parameters
        - outputConfig
    Job:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the job.
          readOnly: true
        projectId:
          type: string
          format: uuid
          description: ID of the project this job belongs to.
          readOnly: true
        customerId:
          type: string
          format: uuid
          description: ID of the customer who owns the job.
          readOnly: true
        name:
          type: string
          description: Optional name for the job.
        description:
          type: string
          description: Optional description for the job.
        status:
          $ref: '#/components/schemas/JobStatusValue'
        progress:
          $ref: '#/components/schemas/JobProgress'
        error:
          $ref: '#/components/schemas/JobError'
        config:
          $ref: '#/components/schemas/JobConfiguration'
        outputUri: # Add direct link to output if available after completion
            type: string
            format: uri
            description: GCS URI of the final generated data output.
            readOnly: true
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the job was created.
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the job was last updated.
          readOnly: true
        startedAt:
          type: string
          format: date-time
          description: Timestamp when the job started processing.
          readOnly: true
        completedAt:
          type: string
          format: date-time
          description: Timestamp when the job finished (completed, failed, or cancelled).
          readOnly: true
      required:
        - id
        - projectId
        - customerId
        - status
        - config
        - createdAt
        - updatedAt
    CreateJobRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        config:
          $ref: '#/components/schemas/JobConfiguration'
      required:
        - name
        - config
...
  securitySchemes:
    BearerAuth: # Or could be cookie-based session
      type: http
      scheme: bearer
      bearerFormat: JWT # Or describe session cookie if used
    SessionCookieAuth: # Changed to reflect actual implementation
      type: apiKey
      in: cookie
      name: session_token # Name of the session cookie used by auth service
  responses: # Defined common responses here
    UnauthorizedError:
      description: Authentication information is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - BearerAuth: [] # Apply security globally, can override per-operation
  - SessionCookieAuth: [] # Apply security globally using session cookies
paths:
...
  /projects/{projectId}/jobs:
    parameters:
      - $ref: '#/components/parameters/ProjectId' # Reference common parameter
    get:
      summary: List data generation jobs for a project
      tags:
        - Jobs
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/JobStatusValue'
          description: Filter jobs by status.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Maximum number of jobs to return.
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Number of jobs to skip for pagination.
      responses:
        '200':
          description: A list of data generation jobs.
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User does not have access to this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new data generation job for a project
      tags:
        - Jobs
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '201': # Or 202 Accepted if job runs async
          description: Job created/accepted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job' # Return the initial job state
        '400':
          description: Invalid input (e.g., invalid config).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User cannot create jobs in this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '404':
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '429':
          description: Rate limit exceeded for creating jobs.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '500':
          $ref: '#/components/responses/InternalServerError'
  /jobs/{jobId}:
    parameters:
      - name: jobId
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: ID of the data generation job.
    get:
      summary: Get data generation job details and status
      tags:
        - Jobs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Job details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User does not have access to this job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '404':
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete: # Using DELETE for cancellation request
      summary: Cancel a running or queued data generation job
      tags:
        - Jobs
      security:
        - BearerAuth: []
      responses:
        '200': # Return updated job status after cancellation request
           description: Job cancellation requested successfully. Returns updated job status.
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/Job'
        '400':
          description: Job cannot be cancelled (e.g., already completed or failed).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User cannot cancel this job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '404':
          description: Job not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse' # Added ErrorResponse ref
        '500':
          $ref: '#/components/responses/InternalServerError'
...